+++ src/lib/items/Item.js
@@ -32,8 +32,10 @@ export default class Item extends Component {
     // onDrop: React.PropTypes.func,
     // onResizing: React.PropTypes.func,
     // onResized: React.PropTypes.func,
-    // onContextMenu: React.PropTypes.func
+    // onContextMenu: React.PropTypes.func,
+    // itemRenderer: React.PropTypes.func
   }
+
   static defaultProps = {
     selected: false
   }
@@ -195,6 +197,7 @@ export default class Item extends Component {
       .draggable({
         enabled: this.props.selected
       })
+      .styleCursor(false)
       .on('dragstart', (e) => {
         if (this.props.selected) {
           this.setState({
@@ -432,6 +435,15 @@ export default class Item extends Component {
     }
   }
 
+  renderContent () {
+    const Comp = this.props.itemRenderer
+    if (Comp) {
+      return <Comp item={this.props.item} />
+    } else {
+      return this.itemTitle
+    }
+  }
+
   render () {
     const dimensions = this.props.dimensions
     if (typeof this.props.order === 'undefined' || this.props.order === null) {
@@ -472,7 +484,7 @@ export default class Item extends Component {
         {this.props.useResizeHandle ? <div ref='dragLeft' className='rct-drag-left'></div> : ''}
         <div className='rct-item-overflow'>
           <div className='rct-item-content'>
-            {this.itemTitle}
+            {this.renderContent()}
           </div>
         </div>
         {this.props.useResizeHandle ? <div ref='dragRight' className='rct-drag-right'></div> : ''}